requirements.txt
streamlit
pypandoc_binary
